requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
whitenoise==6.8.2
django-cors-headers==4.7.0
flake8==7.1.1
vulture==2.14
bandit==1.8.2
safety==3.3.0
black==25.1.0
isort==6.0.0
mypy==1.15.0
django-stubs==5.1.3
openpyxl==3.1.5
